Don't run test on linux - it does not apply
